changed from float to double

+++ stability_condition.cpp
@@ -15,7 +15,7 @@
 
 using namespace std;
 
-float stable_time (double x_extent, double y_extent, int nx, int ny, float kappa,
+double stable_time (double x_extent, double y_extent, int nx, int ny, double kappa,
                     gsl_matrix * vx, gsl_matrix * vy, double dx, double dy) {
 
     double dt;
